Remove `Bag-Size` from example bag-info.txt
This field is not specified but its presence in the examples can cause
confusion that it's a standard field. Since any app which needs a
human-formatted total file size may trivially calculate it from the
Payload-Oxum (also see #2) or filesystem metadata this commit removes it
to avoid confusion or the implication that tools must maintain the
value.