Avoid confusing "sub-directory must not be changed"
It is out of scope for this specification to specify "changes" to bags
or the process of archiving an existing directory file structure into
bagit.

Rather we should say that the payload files may be in sub-directories
(data/ is not flat), and just that the directory and filenames have no
given meaning .. HERE.

Obviously, just like the octet bytes they have a
meaning to someone.

(If we are going to talk about "change" this opens a can of worms about
changes to checksums, changes to tag files, augmenting bag-info.txt etc
etc)